+++ src/rust/jsg-test/tests/mod.rs
@@ -1,0 +1,5 @@
+mod eval;
+mod jsg_struct;
+mod non_coercible;
+mod resource_callback;
+mod unwrap;
