+++ src/rust/jsg-test/tests/non_coercible.rs
@@ -17,50 +17,50 @@ struct MyResource {
 impl MyResource {
     #[jsg_method]
     pub fn string(&self, nc: NonCoercible<String>) -> Result<String, String> {
-        Ok(nc.value().clone())
+        Ok(nc.as_ref().clone())
     }
 
     #[jsg_method]
     pub fn boolean(&self, nc: NonCoercible<bool>) -> Result<bool, String> {
-        Ok(*nc.value())
+        Ok(*nc.as_ref())
     }
 
     #[jsg_method]
     pub fn number(&self, nc: NonCoercible<f64>) -> Result<f64, String> {
-        Ok(*nc.value())
+        Ok(*nc.as_ref())
     }
 }
 
 #[test]
-fn non_coercible_string_new_and_value() {
+fn non_coercible_string_new_and_as_ref() {
     let nc: NonCoercible<String> = NonCoercible::new("hello".to_owned());
-    assert_eq!(nc.value(), "hello");
+    assert_eq!(nc.as_ref(), "hello");
 }
 
 #[test]
-fn non_coercible_bool_new_and_value() {
+fn non_coercible_bool_new_and_as_ref() {
     let nc_true: NonCoercible<bool> = NonCoercible::new(true);
     let nc_false: NonCoercible<bool> = NonCoercible::new(false);
-    assert!(*nc_true.value());
-    assert!(!*nc_false.value());
+    assert!(*nc_true.as_ref());
+    assert!(!*nc_false.as_ref());
 }
 
 #[test]
-fn non_coercible_f64_new_and_value() {
+fn non_coercible_f64_new_and_as_ref() {
     let nc: NonCoercible<f64> = NonCoercible::new(2.5);
-    assert!((*nc.value() - 2.5).abs() < f64::EPSILON);
+    assert!((*nc.as_ref() - 2.5).abs() < f64::EPSILON);
 }
 
 #[test]
 fn non_coercible_from_trait() {
     let nc_string: NonCoercible<String> = "test".to_owned().into();
-    assert_eq!(nc_string.value(), "test");
+    assert_eq!(nc_string.as_ref(), "test");
 
     let nc_bool: NonCoercible<bool> = true.into();
-    assert!(*nc_bool.value());
+    assert!(*nc_bool.as_ref());
 
     let nc_f64: NonCoercible<f64> = 42.0.into();
-    assert!((*nc_f64.value() - 42.0).abs() < f64::EPSILON);
+    assert!((*nc_f64.as_ref() - 42.0).abs() < f64::EPSILON);
 }
 
 #[test]
@@ -94,7 +94,7 @@ fn non_coercible_deref_trait() {
 fn non_coercible_clone() {
     let nc1: NonCoercible<String> = NonCoercible::new("clone test".to_owned());
     let nc2 = nc1.clone();
-    assert_eq!(nc1.value(), nc2.value());
+    assert_eq!(nc1.as_ref(), nc2.as_ref());
     assert_eq!(*nc1, *nc2);
 }
 
@@ -133,7 +133,7 @@ fn non_coercible_methods_accept_correct_types_and_reject_incorrect_types() {
         });
         let mut template = MyResourceTemplate::new(&mut lock);
         let wrapped = jsg::wrap_resource(&mut lock, resource, &mut template);
-        ctx.set_global_safe("resource", wrapped.as_ffi());
+        ctx.set_global_safe("resource", wrapped.into_ffi());
 
         // String method accepts string
         assert_eq!(
